Dwl mode. Added possibility avoiding download projects throwed Transport ex from jgit.
This was achieved with recursion call main method in catch block when jgit throwed Transport exception when some project do not downloading.
Disadventage of this approah in this code  is many times show message "Downloading n projects.".
For exampe if in dataset (source input ODS file) contain 5 projects and 2 of them downloading with error it show message of the end downloading:
Downloaded 5 projects.
Finish.
Downloaded 3 projects.
Finish.